requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cycler==0.10.0
discord.py==1.7.1
dotenv-python==0.0.1
gevent==21.1.2
greenlet==1.0.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.1
multidict==4.7.6
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
praw==7.1.0
prawcore==1.5.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.24.0
six==1.15.0
update-checker==0.18.0
urllib3==1.25.11
websocket-client==0.57.0
yarl==1.5.1
zope.event==4.5.0
zope.interface==5.4.0
